iris-mlops/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iris-mlops"
version = "0.1.0"
description = "Iris classification/regression with MLOps workflow scaffold"
authors = [{ name = "Nikita" }]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "numpy==1.26.4",
    "pandas==2.2.3",
    "scikit-learn==1.5.2",
    "pyyaml==6.0.2",
    "matplotlib==3.10.1",
    "mlflow==2.16.0",
    "wandb==0.23.1",
    "hydra-core==1.3.2",
]

[project.optional-dependencies]
hw4 = [
    "apache-airflow==2.9.3",
]
hw5 = [
    "clearml==2.0.2",
]
docs = [
    "mkdocs==1.6.1",
    "mkdocs-material==9.7.1",
    "pymdown-extensions==10.19.1",
    "tabulate==0.9.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["iris_mlops"]
skip = ["data", "models", "reports", "notebooks", ".venv", "venv"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "B", "I", "S"]
ignore = ["S101"]  # allow assert in tests
extend-exclude = ["data", "models", "reports", "notebooks", ".venv", "venv"]

[tool.ruff.isort]
known-first-party = ["iris_mlops"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = true
strict_optional = true
exclude = "(data|models|reports|notebooks|\\.venv|venv)"

[tool.setuptools.packages.find]
where = ["src"]
